scripts/requirements.txt
matplotlib
numpy
requests
json
opencv
pillow
pandas